refactor: split components into modules
